Added testing script
Added script for testing new plugins. This script will run through calling each of the modules functions, including the main function with and with out arguments, and return a list of errors it encountered. If it returns an empty array the plugin should be ready to add to the bot. This script verifies that the plugin does not cause an exception, it is not designed to test for expected behavior.